Ystrdy record table building correctly

+++ app/src/main/java/com/tobros/hatebyte/ystrdy/weatherrecords/entity/YstrdyRecordEntity.java
@@ -10,7 +10,7 @@ import java.util.Date;
 public class YstrdyRecordEntity {
 
     public float difference;
-    public int nowRecordId;
+    public long nowRecordId;
     public Date date;
     public NowRecordEG nowRecordGateway;
